requirements.txt
# Privacy-First Personal Agent Dependencies
# Core LLM and NLP
ollama>=0.1.0
spacy>=3.7.0
textblob>=0.17.1

# Encryption (removed pysqlcipher3 - Windows compatibility issues)
cryptography>=41.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface (Optional)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
keyring>=24.2.0
colorama>=0.4.6
tabulate>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
